+++ app/src/main/java/com/jennifer/adapter/RecyclerAdapter.java
@@ -2,17 +2,14 @@ package com.jennifer.adapter;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v7.widget.CardView;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.TextView;
 
 import com.jennifer.R;
 import com.jennifer.controller.SecondActivity;
-import com.jennifer.model.Ranking;
 import com.jennifer.model.User;
 
 import java.util.List;
@@ -22,9 +19,9 @@ import java.util.List;
  */
 public class RecyclerAdapter extends RecyclerView.Adapter<RecyclerAdapter.ViewHolder> {
 
-    private List<Ranking> mItems;
+    private List<User> mItems;
 
-    public RecyclerAdapter(List<Ranking> items) {
+    public RecyclerAdapter(List<User> items) {
         mItems = items;
     }
 
@@ -37,7 +34,7 @@ public class RecyclerAdapter extends RecyclerView.Adapter<RecyclerAdapter.ViewHo
 
     @Override
     public void onBindViewHolder(ViewHolder viewHolder, int i) {
-        Ranking item = mItems.get(i);
+        User item = mItems.get(i);
 
         // Data Set
         viewHolder.mImage.setImageResource(item.getImagen());
